+++ src/app/pages/event/view/[username]/[event_name]/attendance.vue
@@ -4,7 +4,6 @@
       v-if="event && route.params.username === store.signedInUsername"
       class="flex flex-col gap-4"
     >
-      <LayoutBreadcrumbs :items="breadcrumbItems" />
       <LayoutPageTitle :title="t('title')" />
       <Steps
         :active="t('qrCodeScan')"
@@ -70,27 +69,13 @@ import {
 import type { RouteLocationNormalized } from 'vue-router'
 import type { RouteNamedMap } from 'vue-router/auto-routes'
 
-import { usePageBreadcrumb as usePageBreadcrumbEvents } from '../../../index.vue'
-import { usePageBreadcrumb as usePageBreadcrumbEventsUser } from '../index.vue'
-import { pageBreadcrumb as usePageBreadcrumbEventsUserId } from './index.vue'
-
 import { useEventByAuthorAccountIdAndSlugQuery } from '~~/gql/documents/queries/event/eventByAuthorAccountIdAndSlug'
 import { getEventItem } from '~~/gql/documents/fragments/eventItem'
 import { useAccountByUsernameQuery } from '~~/gql/documents/queries/account/accountByUsername'
 import { getAccountItem } from '~~/gql/documents/fragments/accountItem'
-import type { BreadcrumbLinkLocalized } from '~/types/breadcrumbs'
 
 const ROUTE_NAME: keyof RouteNamedMap =
   'event-view-username-event_name-attendance___en'
-
-export const usePageBreadcrumb = () => {
-  const route = useRoute(ROUTE_NAME)
-
-  return {
-    label: 'Check-in',
-    to: `/event/view/${route.params.username}/${route.params.event_name}/attendance`,
-  } as BreadcrumbLinkLocalized
-}
 
 setZXingModuleOverrides({
   locateFile: (path, prefix) => {
@@ -119,12 +104,10 @@ definePageMeta({
   },
 })
 
-const { $urql } = useNuxtApp()
 const { t } = useI18n()
 const store = useMaevsiStore()
 const route = useRoute(ROUTE_NAME)
 const fireAlert = useFireAlert()
-const getBreadcrumbItemProps = useGetBreadcrumbItemProps()
 
 // api data
 const accountByUsernameQuery = await zalgo(
@@ -148,15 +131,6 @@ const event = computed(() =>
 const api = getApiData([accountByUsernameQuery, eventQuery])
 
 // data
-const breadcrumbItems = getBreadcrumbItemProps([
-  usePageBreadcrumbEvents(),
-  usePageBreadcrumbEventsUser(),
-  await usePageBreadcrumbEventsUserId({ $urql, route }),
-  {
-    current: true,
-    ...usePageBreadcrumb(),
-  },
-])
 const invitationId = ref<string>()
 const isNfcWritableErrorMessage = ref<string>()
 const loading = ref(true)