+++ src/app/pages/session/view/[id].vue
@@ -1,6 +1,5 @@
 <template>
   <div>
-    <LayoutBreadcrumbs :items="breadcrumbItems" />
     <LayoutPageTitle :title="title" />
     <div class="flex flex-col gap-8">
       <section class="flex flex-col gap-4">
@@ -189,12 +188,11 @@
                 </template>
               </ButtonColored>
               <ButtonColored
-                :aria-label="t('copyFcmToken')"
+                :aria-label="t('showFcmToken')"
                 :is-primary="false"
-                :disabled="!fcmToken"
-                @click="copyFcmToken"
+                @click="showFcmToken"
               >
-                {{ t('copyFcmToken') }}
+                {{ t('showFcmToken') }}
                 <template #prefix>
                   <IHeroiconsClipboard />
                 </template>
@@ -207,45 +205,17 @@
   </div>
 </template>
 
-<script lang="ts">
-import { consola } from 'consola'
-import type { RouteNamedMap } from 'vue-router/auto-routes'
-
-import { usePageBreadcrumb as usePageBreadcrumbHome } from '../../index.vue'
-import type { BreadcrumbLinkLocalized } from '~/types/breadcrumbs'
-
-const ROUTE_NAME: keyof RouteNamedMap = 'session-view-id___en'
-
-export const usePageBreadcrumb = () => {
-  const route = useRoute(ROUTE_NAME)
-
-  return {
-    label: {
-      de: 'Sitzung',
-      en: 'Session',
-    },
-    to: `/session/view/${route.params.id}`,
-  } as BreadcrumbLinkLocalized
-}
-</script>
-
 <script setup lang="ts">
+import { consola } from 'consola'
+
 const { t } = useI18n()
 const requestEvent = useRequestEvent()
-const getBreadcrumbItemProps = useGetBreadcrumbItemProps()
 const store = useMaevsiStore()
 const dateTime = useDateTime()
 const { signOut } = useSignOut()
+const fireAlert = useFireAlert()
 
 // data
-const breadcrumbItems = getBreadcrumbItemProps([
-  usePageBreadcrumbHome(),
-  {
-    current: true,
-    ...usePageBreadcrumb(),
-  },
-])
-const fcmToken = ref<string>()
 const isNavigatorHavingPermissions = ref<boolean>()
 const isNavigatorHavingServiceWorker = ref<boolean>()
 const isWindowHavingNotification = ref<boolean>()
@@ -254,9 +224,6 @@ const permissionState = ref<PermissionState>()
 const title = t('title')
 
 // methods
-const copyFcmToken = async () => {
-  if (fcmToken.value) await copyText(fcmToken.value)
-}
 const sendNotification = async () => {
   const serviceWorkerRegistration = await navigator.serviceWorker.ready
 
@@ -264,6 +231,14 @@ const sendNotification = async () => {
     body: "It's great to see you!",
     icon: '/assets/static/logos/maevsi_icon.svg',
     tag: 'test',
+  })
+}
+const showFcmToken = async () => {
+  const token = await requestFcmToken()
+  await fireAlert({
+    level: 'info',
+    title: 'FCM Token',
+    text: token,
   })
 }
 const requestNotificationPermissions = () =>
@@ -319,17 +294,9 @@ onMounted(async () => {
           permissionStatus.state,
       )
       permissionState.value = permissionStatus.state
-
-      if (permissionStatus.state === 'granted') {
-        fcmToken.value = await requestFcmToken()
-      }
     })
 
     permissionState.value = permissionStatus.state
-
-    if (permissionStatus.state === 'granted') {
-      fcmToken.value = await requestFcmToken()
-    }
   }
 
   if (!permissionState.value && isWindowHavingNotification.value) {
@@ -359,7 +326,7 @@ de:
   notificationSend: Benachrichtigung senden
   sessionExpiry: Deine Sitzung läuft am {exp} ab.
   sessionExpiryNone: Es sind keine Sitzungsdaten verfügbar.
-  copyFcmToken: FCM Token kopieren
+  showFcmToken: FCM Token anzeigen
   title: Sitzung
   userAgentString: User agent string
 en:
@@ -378,7 +345,7 @@ en:
   notificationSend: Send notification
   sessionExpiry: Your session expires on {exp}.
   sessionExpiryNone: No session data is available.
-  copyFcmToken: Copy FCM Token
+  showFcmToken: Show FCM Token
   title: Session
   userAgentString: User agent string
 </i18n>