+++ src/tailwind.config.ts
@@ -1,3 +1,5 @@
+import forms from '@tailwindcss/forms'
+import typography from '@tailwindcss/typography'
 import type { Config } from 'tailwindcss'
 import defaultTheme from 'tailwindcss/defaultTheme'
 import colors from 'tailwindcss/colors'
@@ -69,6 +71,8 @@ export default <Partial<Config>>{
         },
       })
     },
+    forms(),
+    typography(),
   ],
   theme: {
     extend: {