+++ tests/e2e/specs/pages/account/accountUsername.spec.ts
@@ -34,34 +34,6 @@ test.describe('account username page', () => {
   // })
   //
   // TODO: mock data
-  // TODO: add breadcrumb tests for other pages
-  // context('navigation', () => {
-  //   it('has working breadcrumbs', () => {
-  //     cy.visit({
-  //       url: '/accounts/dargmuesli',
-  //     })
-  //     cy.get('[data-testid="breadcrumb-prefix-/"]').click()
-  //     cy.url().should('equal', `${CYPRESS_BASE_URL(Cypress.env('NODE_ENV'))}/`)
-  //     cy.visit({
-  //       url: '/accounts/dargmuesli',
-  //     })
-  //     cy.get('[data-testid="breadcrumb-prefix-/accounts"]').click()
-  //     cy.url().should(
-  //       'equal',
-  //       `${CYPRESS_BASE_URL(Cypress.env('NODE_ENV'))}/tasks/accounts/sign-in`
-  //     )
-  //     cy.visit({
-  //       url: '/accounts/dargmuesli',
-  //     })
-  //     cy.get('[data-testid="breadcrumb-?"]').click()
-  //     cy.url().should(
-  //       'equal',
-  //       `${CYPRESS_BASE_URL(Cypress.env('NODE_ENV'))}/accounts/dargmuesli`
-  //     )
-  //   })
-  // })
-  //
-  // TODO: mock data
   // context('visual regression', () => {
   //   it('looks as before', () => {
   //     cy.visit('/accounts/username')